==> notebooks/labr_notebooks/00_data_loader.ipynb <==
import pandas as pd

# clone the LABR repo from github
file_path = "../data/LABR/data/reviews.tsv"

df = pd.read_csv(file_path, sep="\t", header=None)
df.columns = ["rating", "review_id", "user_id", "book_id", "review_text"]
print(f"Dataset shape: {df.shape}")
print(df.head())

for col in df.columns:
    print(col)

# save dataset
df.to_csv('./row/labr_row.csv')
