==> notebooks/labr_notebooks/03_camel_bert_sentiment.ipynb <==
import pandas as pd
import torch
from transformers import AutoTokenizer, AutoModelForSequenceClassification
from torch.utils.data import Dataset, DataLoader

df = pd.read_csv("../data/processed/labr_cleaned.csv") 
texts = df["review_text_clean"].fillna("").tolist()

MODEL_NAME = "CAMeL-Lab/bert-base-arabic-camelbert-da-sentiment"
tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
model = AutoModelForSequenceClassification.from_pretrained(MODEL_NAME)
model.eval()

torch.set_num_threads(12)
print(f"Using {torch.get_num_threads()} threads for inference")

label_map_index = {0: -1, 1: 0, 2: 1}

class ReviewDataset(Dataset):
    def __init__(self, texts, tokenizer, max_length=128):
        self.texts = texts
        self.tokenizer = tokenizer
        self.max_length = max_length

    def __len__(self):
        return len(self.texts)

    def __getitem__(self, idx):
        text = self.texts[idx]
        encoding = self.tokenizer(
            text,
            truncation=True,
            padding="max_length",
            max_length=self.max_length,
            return_tensors="pt"
        )
        return {k: v.squeeze(0) for k, v in encoding.items()}

dataset = ReviewDataset(texts, tokenizer)
dataloader = DataLoader(dataset, batch_size=16, shuffle=False, num_workers=10)


all_labels = []

with torch.no_grad():
    for batch in dataloader:
        outputs = model(**batch)
        preds = torch.argmax(torch.softmax(outputs.logits, dim=1), dim=1)
        all_labels.extend([label_map_index[p.item()] for p in preds])

# =========================
# 7. Save results
# =========================
df["sentiment_label"] = all_labels
df.to_csv("../data/processed/labr_cleaned_sentiment.csv", index=False)
