==> notebooks/labr_notebooks/04_rating_based_sentiment_labling.ipynb <==
# LABR dataset authors's method
# Mohamed Aly and Amir Atiya

import pandas as pd
import numpy as np

df = pd.read_csv("../data/processed/labr_cleaned_sentiment_camelbert.csv")
print("Dataset shape:", df.shape)
df.head(2)

print("Columns after cleanup:")
print(df.columns)


df = df[[
'rating',
'review_id', 
'user_id',
'book_id',
'review_text',
'review_text_clean',
'text_length',
'sentiment_label'
]]
print(df.columns)

def rating_to_sentiment(r):
    if r in [4, 5]:
        return 1
    elif r == 3:
        return 0
    elif r in [1, 2]:
        return -1
    else:
        return np.nan
df["ground_truth"] = df["rating"].apply(rating_to_sentiment)

df.head(2)

df.to_csv('../data/processed/labr_cleaned_sentiment_camelbert_and_author_score.csv')
