==> notebooks/labr_notebooks/05_calculate_final_score.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('../data/processed/labr_cleaned_sentiment_camelbert_and_author_score.csv')
print("Dataset shape:", df.shape)
df.head(2)

df = df.loc[:, ~df.columns.str.contains("^Unnamed")]
print("Columns after cleanup:")
print(df.columns)

# ground_truth: -1 -> 0, 0 -> 0.5, 1 -> 1
df["sentiment_score_normalized"] = (df["ground_truth"] + 1) / 2

max_rating = 5
df["score_normalized"] = df["rating"] / max_rating

df["final_score"] = (
    0.5 * df["score_normalized"] + 0.5 * df["sentiment_score_normalized"]
) * max_rating

df.columns

# save result
df.to_csv('../data/processed/labr_complete.csv')
